combined farmer actions and user actions into single file for each
